layouts fixed + profile page removed
