+++ src/pages/NewsfeedPage/NewsfeedPage.jsx
@@ -1,18 +1,10 @@
 import React, { useState, useEffect } from "react";
-import { useQuery, useSubscription } from "@apollo/react-hooks";
-import { notification, Row } from "antd";
+import { useQuery } from "@apollo/react-hooks";
+import { Row } from "antd";
 import Navbar from "../../components/Navbar/Navbar";
 import SingleChat from "../../components/Message/SingleChat";
-import Notification from "../../components/Notification/Notification";
 import Post from "../../components/Post/Post";
-import {
-  LOAD_USER,
-  NEW_NOTIFICATION,
-  DELETE_NOTIFICATION,
-  GET_NOTIFICATIONS,
-  GET_NEWSFEED,
-  NEW_POST
-} from "../../utils/queries";
+import { LOAD_USER, GET_NEWSFEED, NEW_POST } from "../../utils/queries";
 import {
   InfoContainer,
   PostsSection,
@@ -33,9 +25,6 @@ const NewsfeedPage = () => {
     visible: false,
     creator: null
   });
-
-  useQuery(GET_NOTIFICATIONS);
-
   const { data: newsfeedData, subscribeToMore } = useQuery(GET_NEWSFEED);
 
   useEffect(() => {
@@ -48,64 +37,6 @@ const NewsfeedPage = () => {
       }
     });
   }, [subscribeToMore]);
-
-  const openNotification = newNotification => {
-    notification.info({
-      message: <span style={{ marginLeft: "40px" }}>Notification</span>,
-      description: (
-        <div style={{ marginTop: "12px" }}>
-          <Notification notification={newNotification} />
-        </div>
-      ),
-      duration: 0,
-      placement: "bottomLeft",
-      style: {
-        padding: "16px 8px"
-      }
-    });
-  };
-
-  useSubscription(NEW_NOTIFICATION, {
-    onSubscriptionData: ({ client, subscriptionData }) => {
-      const data = client.readQuery({
-        query: GET_NOTIFICATIONS
-      });
-      if (
-        subscriptionData.data.newNotification.notifier.id ===
-        userData.loadUser.id
-      ) {
-        openNotification(subscriptionData.data.newNotification);
-        const newData = {
-          getNotifications: [
-            subscriptionData.data.newNotification,
-            ...data.getNotifications
-          ]
-        };
-
-        client.writeQuery({
-          query: GET_NOTIFICATIONS,
-          data: newData
-        });
-      }
-    }
-  });
-
-  useSubscription(DELETE_NOTIFICATION, {
-    onSubscriptionData: ({ client, subscriptionData }) => {
-      const data = client.readQuery({
-        query: GET_NOTIFICATIONS
-      });
-
-      const newNotificationList = data.getNotifications.filter(
-        item => item.id !== subscriptionData.data.deleteNotification
-      );
-
-      client.writeQuery({
-        query: GET_NOTIFICATIONS,
-        data: { getNotifications: newNotificationList }
-      });
-    }
-  });
 
   return (
     <>
