+++ src/pages/ProfilePage/ProfilePage.jsx
@@ -1,20 +1,14 @@
 import React, { useState } from "react";
 import { useParams } from "react-router-dom";
-import { useQuery, useSubscription } from "@apollo/react-hooks";
-import { notification, Row } from "antd";
-import Navbar from "../../components/Navbar/Navbar";
+import { useQuery } from "@apollo/react-hooks";
+import { Row } from "antd";
 import SingleChat from "../../components/Message/SingleChat";
-import Notification from "../../components/Notification/Notification";
-import ProfileHeader from "../../components/ProfileHeader/ProfileHeader";
 import CreatePostDefault from "../../components/Post/CreatePostDefault";
 import Post from "../../components/Post/Post";
 import About from "../../components/About/About";
 import {
   LOAD_USER,
   GET_POSTS,
-  NEW_NOTIFICATION,
-  DELETE_NOTIFICATION,
-  GET_NOTIFICATIONS,
   GET_URL_POSTS,
   LOAD_FROM_URL_USER
 } from "../../utils/queries";
@@ -49,67 +43,8 @@ const ProfilePage = () => {
   const { data: profileData } = useQuery(LOAD_FROM_URL_USER, {
     variables: {
       username
-    }
-  });
-
-  useQuery(GET_NOTIFICATIONS);
-
-  const openNotification = newNotification => {
-    notification.info({
-      message: <span style={{ marginLeft: "40px" }}>Notification</span>,
-      description: (
-        <div style={{ marginTop: "12px" }}>
-          <Notification notification={newNotification} />
-        </div>
-      ),
-      duration: 0,
-      placement: "bottomLeft",
-      style: {
-        padding: "16px 8px"
-      }
-    });
-  };
-
-  useSubscription(NEW_NOTIFICATION, {
-    onSubscriptionData: ({ client, subscriptionData }) => {
-      const data = client.readQuery({
-        query: GET_NOTIFICATIONS
-      });
-      if (
-        subscriptionData.data.newNotification.notifier.id ===
-        userData.loadUser.id
-      ) {
-        openNotification(subscriptionData.data.newNotification);
-        const newData = {
-          getNotifications: [
-            subscriptionData.data.newNotification,
-            ...data.getNotifications
-          ]
-        };
-
-        client.writeQuery({
-          query: GET_NOTIFICATIONS,
-          data: newData
-        });
-      }
-    }
-  });
-
-  useSubscription(DELETE_NOTIFICATION, {
-    onSubscriptionData: ({ client, subscriptionData }) => {
-      const data = client.readQuery({
-        query: GET_NOTIFICATIONS
-      });
-
-      const newNotificationList = data.getNotifications.filter(
-        item => item.id !== subscriptionData.data.deleteNotification
-      );
-
-      client.writeQuery({
-        query: GET_NOTIFICATIONS,
-        data: { getNotifications: newNotificationList }
-      });
-    }
+    },
+    skip: !readOnly()
   });
 
   const { data: postsData } = useQuery(GET_POSTS, {
@@ -127,17 +62,6 @@ const ProfilePage = () => {
     <>
       {userData && (
         <Row>
-          <Navbar
-            onProfile
-            user={userData.loadUser}
-            setOpenChat={setOpenChat}
-          />
-          <ProfileHeader
-            user={profileData ? profileData.loadFromUrlUser : userData.loadUser}
-            authUser={userData.loadUser}
-            readOnly={readOnly()}
-            setOpenChat={setOpenChat}
-          />
           <InfoContainer>
             <FixedContainer>
               <About
